+++ Packages/com.github.kurotu.vrc-quest-tools/Runtime/Components/NetworkIDAssigner.cs
@@ -10,7 +10,7 @@ namespace KRT.VRCQuestTools.Components
     [AddComponentMenu("VRCQuestTools/VQT Network ID Assigner")]
     [HelpURL("https://kurotu.github.io/VRCQuestTools/docs/references/components/network-id-assigner?lang=auto")]
     [DisallowMultipleComponent]
-    public class NetworkIDAssigner : VRCQuestToolsEditorOnly, INdmfComponent
+    public class NetworkIDAssigner : VRCQuestToolsEditorOnly
     {
         // no members.
     }